// part9/flight-diary/frontend/flight-dairy/src/types.ts
export interface Flight {
  id: string;
  date: string;
  weather: string;
  visibility: string;
  comment: string;
}

export enum Weather {
  Sunny = 'sunny',
  Rainy = 'rainy',
  Cloudy = 'cloudy',
  Stormy = 'stormy',
  Windy = 'windy',
}

export enum Visibility {
  Great = 'great',
  Good = 'good',
  Ok = 'ok',
  Poor = 'poor',
}